+++ marshmallow/__init__.py
@@ -7,7 +7,7 @@ from marshmallow.schema import (
     MarshalResult,
     UnmarshalResult,
 )
-from marshmallow.decorators import pre_dump, post_dump, pre_load, post_load
+from marshmallow.decorators import pre_dump, post_dump, pre_load, post_load, validator
 from marshmallow.utils import pprint, missing
 from marshmallow.exceptions import MarshallingError, UnmarshallingError, ValidationError
 
@@ -18,6 +18,7 @@ __license__ = 'MIT'
 __all__ = [
     'Schema',
     'SchemaOpts',
+    'validator',
     'pre_dump',
     'post_dump',
     'pre_load',